feat: add README, linter
