+++ runner
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+go run cmd/main.go $1 $2
